+++ languagetool-language-modules/ar/src/main/java/org/languagetool/language/Arabic.java
@@ -127,6 +127,9 @@ public class Arabic extends LanguageWithModel {
       new ArabicWordinessRule(messages),
       new ArabicWrongWordInContextRule(messages, this),
       new ArabicTransVerbRule(messages),
+      new ArabicTransVerbDirectToIndirectRule(messages),
+      new ArabicTransVerbIndirectToDirectRule(messages),
+      new ArabicTransVerbIndirectToIndirectRule(messages),
       new ArabicInflectedOneWordReplaceRule(messages)
     );
   }