+++ languagetool-language-modules/ar/src/main/java/org/languagetool/tagging/ar/ArabicTagManager.java
@@ -313,7 +313,7 @@ public class ArabicTagManager {
    * @return test if word has stopword tagging
    */
   public boolean isStopWord(String postag) {
-    return postag.startsWith("P");
+    return (postag != null && postag.startsWith("P"));
   }
 
   /**
@@ -350,6 +350,7 @@ public class ArabicTagManager {
   public boolean isMasdar(String postag) {
     return postag != null && postag.startsWith("NM");
   }
+
 
   /**
    * @return true if have flag is noun and definite